DEV-25:fixed comment and create account & register for google account
