+++ package.json
@@ -6,18 +6,15 @@
   "dependencies": {
     "@ant-design/icons": "^4.6.2",
     "@reduxjs/toolkit": "^1.6.1",
+    "@storybook/addon-actions": "^6.3.6",
+    "@storybook/addon-essentials": "^6.3.6",
+    "@storybook/addon-links": "^6.3.6",
+    "@storybook/node-logger": "^6.3.6",
+    "@storybook/preset-create-react-app": "^3.2.0",
+    "@storybook/react": "^6.3.6",
     "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
     "antd": "^4.16.8",
-    "core-js": "^3.16.0",
-    "core-js-compat": "^3.16.0",
-    "core-js-pure": "^3.15.0",
-    "lint-staged": "^11.1.2",
     "node-sass": "^6.0.1",
-    "object-inspect": "^1.10.3",
-    "rc-image": "^5.2.4",
-    "rc-pagination": "^3.1.6",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-facebook-login": "^4.1.1",
@@ -35,12 +32,6 @@
     "web-vitals": "^1.1.2"
   },
   "devDependencies": {
-    "@storybook/addon-actions": "^6.3.6",
-    "@storybook/addon-essentials": "^6.3.6",
-    "@storybook/addon-links": "^6.3.6",
-    "@storybook/node-logger": "^6.3.6",
-    "@storybook/preset-create-react-app": "^3.2.0",
-    "@storybook/react": "^6.3.6",
     "@types/jest": "^26.0.24",
     "@types/node": "^16.3.3",
     "@types/react": "^17.0.14",
@@ -54,9 +45,7 @@
     "eslint-config-airbnb": "^18.2.1",
     "eslint-config-prettier": "^8.3.0",
     "eslint-config-react-app": "^6.0.0",
-    "eslint-plugin-flowtype": "^5.9.0",
     "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-jest": "^24.4.0",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-react": "^7.24.0",
     "eslint-plugin-react-hooks": "^4.2.0",
