+++ src/pages/SignUp/SignUp.scss
@@ -6,7 +6,9 @@ $width: 100%;
   height: 100vh;
   .img-block {
     border-radius: $border-size;
-
+background-image: url('https://source.unsplash.com/Mv9hjnEUHR4/600x800');
+background-position: 25% 100%;
+background-repeat: no-repeat;
     img {
       width: $width;
       height: 100%;
@@ -41,9 +43,13 @@ $width: 100%;
         background-color: lightgray;
         border: none;
       }
-.email,.last-name{
-  height: 39px;
-}
+      .first-name{
+        height: 39px;
+        border-radius: 20px;
+      }
+      .email,.last-name{
+        height: 39px;
+      }
       .email,
       .password,.last-name {
         border-radius: 20px;
@@ -53,9 +59,9 @@ $width: 100%;
         height: 39px;
         border-radius: 20px;
       }
-.email-form{
-  margin-bottom: 25px;
-}
+      .email-form{
+        margin-bottom: 25px;
+      }
       .ant-btn-dangerous {
         background-color: #f5222d;
         color: white;
@@ -86,6 +92,9 @@ $width: 100%;
 
         i {
           margin-right: 5px;
+        }
+        svg{
+          font-size: 14px;
         }
       }
